pyproject.toml
[tool.poetry]
name = "teste-greenole"
version = "0.1.0"
description = ""
authors = ["carvalhocaio <caiocarvalho.py@gmail.com>"]
readme = "README.md"
packages = [{include = "teste_greenole"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
djangorestframework = "^3.14.0"
django-filter = "^23.3"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
